Backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
